requirements.txt
inject==3.3.2
openpyxl==2.6.2
Pillow==6.2.1
robotframework-seleniumlibrary==4.1.0
